Created Report and fixed bug that would prevent split bets on 0 and 00
Updated how buttons are disabled when placing a split bet to allow the user to place split bets on 0 and 00. Future work may be required to select 0 or 00 as the second number for the split bet.

Added EnableNumericButtons() method for re-enabling buttons after a split bet has been placed.

Created write up about the project documenting how everything works.